Enable building and running web apps using Wasm multi-threading (#499)
# Objective

Closes #226.

Currently, all web apps are running on a single thread, which leaves a
lot of performance on the table.
One of Bevy's premises is that it makes multi-threading automatic to a
large degree, but this is currently not being used for web apps.

Nowadays, multi-threading a Wasm application is possible, but requires a
bunch of configuration.
We want to make it as easy as possible to play around with
multi-threading, so that eventually we can get Bevy to support
multi-threading on the web natively.

# Solution

You can now build and run apps that use multi-threaded Wasm:

- In the CLI: `--unstable multi-threading` or `-U multi-threading`
- In `Config.toml`:
    ```toml
    [package.metadata.bevy_cli.unstable]
    web-multi-threading = true
    ```

Enabling it will do the following:

- Add `-C target-feature=+atomics,+bulk-memory` to the `RUSTFLAGS` env
variable passed to `cargo`. This will enable the Wasm features required
for multi-threading.
- Add `-Z build-std=std,panic_abort` to `cargo build` (only if targeting
`wasm32-unknown-unknown`, `wasm32v1-none` doesn't need std). This will
also enable the Wasm features for std.
- Add the headers `Cross-Origin-Opener-Policy: same-origin` and
`Cross-Origin-Embedder-Policy: require-corp` to all responses of the web
dev server. This enables cross-origin isolation, which is required to
use shared memory.

This requires a nightly Rust toolchain, which is not enforced /
supported by the CLI at the moment to keep the scope smaller.
If the user is not using nightly, they get a fairly useful error message
though:

```
error: the `-Z` flag is only accepted on the nightly channel of Cargo, but this is the `stable` channel
See https://doc.rust-lang.org/book/appendix-07-nightly-rust.html for more information about Rust release channels.
```

The multi-threading flag is marked as experimental and hidden behind the
`unstable` feature, because it requires nightly Rust and also the user
to manually set up their Bevy app to be multi-threaded.

# Testing

To test this, you need to install the CLI from this branch with the
`unstable` feature flag enabled:

```sh
cargo install --git https://github.com/TheBevyFlock/bevy_cli --branch 226-multi-threaded-web-apps -F unstable --locked bevy_cli
```

Then, you need a Bevy app which uses Wasm multi-threading.
Note that Bevy doesn't support this natively yet, so it requires a
custom implementation.

In that app, you can then use `bevy run web -U multi-threading`.
Alternatively you can also configure it in `Cargo.toml`:

```toml
[package.metadata.bevy_cli.unstable]
web-multi-threading = true
```

and then just use `bevy run web`.

To verify the cross-origin isolation, go to the browser dev console and
type `window.crossOriginIsolated`. It should return `true`.


## Example:
[chain-reaction](https://github.com/void-scape/chain-reaction)

@CorvusPrudens gave us a test game that uses multi-threading for audio.

To test it:

```
git clone --recurse-submodules git@github.com:void-scape/chain-reaction.git
```

(It uses submodules, so don't forget the `--recurse-submodules` flag).

Then run

```
bevy run --release web -U multi-threading
```

Check that the soundtrack plays and also check
`window.crossOriginIsolated` in the browser dev console.