Improve docs with undocumented features (#598)
This is a sister PR to #581. Closes #578, closes #479.

I went through each entry in [the
changelog](https://github.com/TheBevyFlock/bevy_cli/blob/update-cli-changelog/CHANGELOG.md)
and, if it was a feature or important detail, ensured it was mentioned
somewhere in our docs. This was one of the last things I wanted to do
before we release!

---------

Co-authored-by: DAA <42379074+DaAlbrecht@users.noreply.github.com>