feat: resolve package according to cargo
Cargo does the following for `build` / `check`:

https://doc.rust-lang.org/cargo/commands/cargo-build.html
> By default, when no package selection options are given, the packages selected depend on the selected manifest file (based on the current working directory if --manifest-path is not given). If the manifest is the root of a workspace then the workspaces default members are selected, otherwise only the package defined by the manifest will be selected.

https://doc.rust-lang.org/cargo/commands/cargo-run.html
When no target selection options are given, cargo run will run the binary target. If there are multiple binary targets, you must pass a target flag to choose one. Or, the default-run field may be specified in the [package] section of Cargo.toml to choose the name of the binary to run by default.