Do not continue `duplicate_bevy_dependencies` if `bevy` is not a dependency (#637)
At the very beginning of the `duplicate_bevy_dependencies` lint, we
count how many crates named `bevy` are available. We currently exit
early if there is only one `bevy` crate, as there cannot be duplicates
in that case.

However, we don't exit early if there are _zero_ `bevy` crates! This
means we're doing a bunch of extra work when there is no `bevy` crate,
just to immediately discard it. This PR fixes the lint to exit early if
there is no crate named `bevy`.

---------

Co-authored-by: DAA <42379074+DaAlbrecht@users.noreply.github.com>