feat: query host tuple through `rustc`, rather than leaving it blank
`ui_test` isn't allowed to find the host tuple itself, since it tries running `bevy_lint_driver -vV` and gets very confused when `bevy_lint_driver` silently eats the `-vV` flag because it expected `rustc`'s path instead.

Because of that, we need to specify the host tuple with some default. Before, leaving it as an empty string worked fine, but in a future commit I'm going to introduce `DependencyBuilder`, which needs a valid host tuple. As such, the best solution is to properly query `rustc` for the host tuple and pass it to `ui_test`.