Implement a cooldown when updating dependencies (#681)
Inspired by [this
article](https://blog.yossarian.net/2025/11/21/We-should-all-be-using-dependency-cooldowns).

This PR configures Dependabot to implement a cooldown which prevents it
from updating dependencies if the new version if less than 7 days old.
The intention is that this will help defend against supply-chain
attacks, in the hope that malignant releases are caught and yanked
before the 7-day cooldown finishes. This cooldown does not apply to
security updates, so we'll still see CVE advisories in the Security tab.